alot have changed acording to new changes from customer

+++ lib/app/modules/categories/category_controller.dart
@@ -1,12 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:vgsync_frontend/app/controllers/global_controller.dart';
+import 'package:vgsync_frontend/app/modules/stock/stock_controller.dart';
+import 'package:vgsync_frontend/app/wigdets/common_widgets.dart';
+import 'package:vgsync_frontend/app/wigdets/custom_notification.dart';
 import '../../data/models/category_model.dart';
 import '../../data/repositories/category_repository.dart';
 
 class CategoryController extends GetxController {
   final CategoryRepository categoryRepository;
   final GlobalController globalController = Get.find<GlobalController>();
+  final StockController stockController = Get.find<StockController>();
 
   CategoryController({required this.categoryRepository});
 
@@ -15,6 +19,8 @@ class CategoryController extends GetxController {
   var isSaving = false.obs;
 
   final nameController = TextEditingController();
+
+  final searchController = TextEditingController();
 
   @override
   void onReady() {
@@ -46,9 +52,16 @@ class CategoryController extends GetxController {
       final added = await categoryRepository.addCategory(newCategory);
       categories.add(added); // reactive update
       fetchCategories();
-      // globalController.triggerRefresh(); // ✅ WRITE event
+
+      stockController.fetchStocks();
+      globalController
+          .triggerRefresh(DashboardRefreshType.stock); // ✅ WRITE event
+
+      Get.back(closeOverlays: true);
+       DesktopToast.show('Category added successfully',  backgroundColor: Colors.greenAccent,);
+  
+
       clearForm();
-      Get.back();
     } finally {
       isSaving.value = false;
     }
@@ -67,20 +80,39 @@ class CategoryController extends GetxController {
       if (index != -1) categories[index] = res; // reactive update
 
       fetchCategories();
-      // globalController.triggerRefresh(); // ✅ WRITE event
+      stockController.fetchStocks();
+      globalController.triggerRefresh(DashboardRefreshType.stock);
 
       clearForm();
-      Get.back();
+
+      Get.back(closeOverlays: true);
+      DesktopToast.show('Category updated successfully',  backgroundColor: Colors.greenAccent,);
+   
     } finally {
       isSaving.value = false;
     }
   }
 
   Future<void> delete(int id) async {
-    await categoryRepository.deleteCategory(id);
-    categories.removeWhere((c) => c.id == id); // reactive update
-    fetchCategories();
-    // globalController.triggerRefresh(); // ✅ WRITE event
+    ConfirmDialog.show(
+      Get.context!,
+      title: "Delete Stock",
+      message: "Are you sure you want to delete this stock item?",
+      onConfirm: () async {
+        await categoryRepository.deleteCategory(id);
+        categories.removeWhere((c) => c.id == id);
+
+        globalController.triggerRefresh(DashboardRefreshType.stock);
+
+        Get.back(closeOverlays: true);
+         DesktopToast.show("Category deleted successfully.",  backgroundColor: Colors.greenAccent,);
+       
+      },
+      confirmText: "Delete",
+      cancelText: "Cancel",
+      snackbarColor: Colors.green,
+      snackbarIcon: Icons.check_circle,
+    );
   }
 
   void fillForm(CategoryModel category) {
@@ -89,5 +121,13 @@ class CategoryController extends GetxController {
 
   void clearForm() {
     nameController.clear();
+  }
+
+  Future<void> refreshCategories() async {
+    // 🔥 RESET FILTERS
+
+    searchController.clear();
+
+    await fetchCategories();
   }
 }
